Complete overhaul of the derive system.
A few breaking changes here:
- indexes are now declared on the model level, not per field.
- indexes are decalred simply as `#[model(index(keys="...",
options="..."))]`. The options field is optional, and may be omitted.
- `skip_serde_checks` is now simply a flag. No boolean values.

In addition to the above, there are fundamental changes in the
underlying mongodb crate. This has necessitated a few fundamental
changes in the attributes which can be derived on models. See the docs
for more details.